refact: 更新ボタンのテストを別ファイルに切り出した
