+++ src/components/BookmarkForm.update-bookmark.test.tsx
@@ -1,0 +1,85 @@
+import { describe, expect, it, vi } from 'vitest'
+import { INVALID_STRING, TestBookmarks } from '../../functions/test/fixtures'
+import userEvent from '@testing-library/user-event'
+import { render, screen } from '@testing-library/react'
+import { BookmarkForm } from './BookmarkForm'
+import { inputText } from '../test/test-utils'
+import { UI_LABELS } from '../../shared/constants/uiMessages'
+
+vi.mock('@tanstack/react-router', () => ({
+  useRouter: () => ({ history: { back: vi.fn() }, navigate: vi.fn() }),
+}))
+
+vi.mock('../hooks/useDeleteBookmark', () => ({
+  useDeleteBookmark: () => ({ mutate: vi.fn(), isPending: false }),
+}))
+
+const mockUpdate = vi.fn()
+const mockIsUpdatePending = false
+
+vi.mock('../hooks/useUpdateBookmark', () => ({
+  useUpdateBookmark: () => ({
+    mutate: mockUpdate,
+    isPending: mockIsUpdatePending,
+  }),
+}))
+
+describe('更新ボタンの動作', () => {
+  const testBookmark = TestBookmarks[0]
+
+  const testData: { name: string; title?: string; url?: string }[] = [
+    { name: 'タイトルもurlも変更しなければ' },
+    { name: 'タイトルが不正ならば', title: '' },
+    { name: 'urlが不正ならば', url: INVALID_STRING.URL },
+    { name: 'urlがftpならば', url: INVALID_STRING.FTP },
+  ]
+
+  it.each(testData)(
+    `$name更新ボタンが無効であること`,
+    async ({ title, url }) => {
+      const user = userEvent.setup()
+
+      render(<BookmarkForm bookmark={testBookmark} />)
+
+      await inputText(user, UI_LABELS.FIELDS.TITLE, title)
+      await inputText(user, UI_LABELS.FIELDS.URL, url)
+
+      const updateButton = await screen.findByRole('button', {
+        name: UI_LABELS.ACTIONS.UPDATE,
+      })
+      expect(updateButton).toBeDisabled()
+    },
+  )
+
+  it('タイトルを正しく変更したら更新ボタンが有効になること', async () => {
+    const user = userEvent.setup()
+
+    render(<BookmarkForm bookmark={testBookmark} />)
+
+    await inputText(user, UI_LABELS.FIELDS.TITLE, TestBookmarks[1].title)
+
+    const updateButton = await screen.findByRole('button', {
+      name: UI_LABELS.ACTIONS.UPDATE,
+    })
+    expect(updateButton).toBeEnabled()
+  })
+
+  it('更新ボタンをクリックしたらupdateBookmark()が呼び出されること', async () => {
+    const user = userEvent.setup()
+    render(<BookmarkForm bookmark={testBookmark} />)
+
+    await inputText(user, UI_LABELS.FIELDS.TITLE, TestBookmarks[1].title)
+
+    const updateButton = await screen.findByRole('button', {
+      name: UI_LABELS.ACTIONS.UPDATE,
+    })
+    await user.click(updateButton)
+
+    // 検証: フックの削除関数が、正しいIDを引数にして呼ばれたか
+    expect(mockUpdate).toHaveBeenCalledWith({
+      id: testBookmark.id,
+      title: TestBookmarks[1].title,
+      url: testBookmark.url,
+    })
+  })
+})
